+++ test/acceptance/app.test.js
@@ -882,7 +882,8 @@ test('field names and types are exposed', function(done){
         url: '/api/v1/sql?' + querystring.stringify({
           q: "SELECT 1::int as a, 2::float8 as b, 3::varchar as c, " +
              "4::char as d, now() as e, 'a'::text as f, " +
-              "'POINT(0 0)'::geometry as the_geom " +
+             "'POINT(0 0)'::geometry as the_geom " +
+             ", 1::bool as g " +
              "LIMIT 0"
         }),
         headers: {host: 'vizzuality.cartodb.com'},
@@ -890,13 +891,14 @@ test('field names and types are exposed', function(done){
     },{ }, function(res) {
         assert.equal(res.statusCode, 200, res.body);
         var parsedBody = JSON.parse(res.body);
-        assert.equal(_.keys(parsedBody.fields).length, 7);
+        assert.equal(_.keys(parsedBody.fields).length, 8);
         assert.equal(parsedBody.fields.a.type, 'number');
         assert.equal(parsedBody.fields.b.type, 'number');
         assert.equal(parsedBody.fields.c.type, 'string');
         assert.equal(parsedBody.fields.d.type, 'string');
         assert.equal(parsedBody.fields.e.type, 'date');
         assert.equal(parsedBody.fields.f.type, 'string');
+        assert.equal(parsedBody.fields.g.type, 'boolean');
         assert.equal(parsedBody.fields.the_geom.type, 'geometry');
         done();
     });